chore: move store modules to sub directory
